backend/requirements.txt
flask==3.1.0
flask-cors==5.0.1
pandas==2.2.3
google-generativeai==0.8.4
python-dotenv==1.1.0
